replication: Drop engine V1 logic
